+++ release.sh
@@ -1,5 +1,5 @@
-#!/bin/bash -x
-export TAG=v3.0.2-rc03
+#!/bin/bash
+export TAG=v3.0.2-rc04
 git tag -d ${TAG}
 git push origin --delete ${TAG}
 gh release delete ${TAG} --repo owlcms/owlcms-controlpanel --yes