+++ ReleaseNotes.md
@@ -10,6 +10,7 @@ The control panel is installed once. It will automatically download the current 
 ## New in version 3.0.2
 - The latest available LTS (Java 25) for the architecture will be installed and used to benefit from performance improvements.
 - An env.properties file is created in each owlcms installation, initialized with the parent env.properties. This allows overriding environment variables per installation.
+- Under Windows, logging control-panel.log in the Control Panel installation directory should now work.
 - owlcms-firmata launching fixed to also use Java 25
 - owlcms-firmata locks released if the control panel is closed by a signal
 - Factually accurate progress bar for the long unzip in tracker